django/requirements.txt
psycopg2-binary
django
djangorestframework
django-cors-headers
daphne
requests
python-dotenv
django-crispy-forms
crispy-bootstrap5
Pillow
django-render-block
channels

argon2-cffi
hvac

django-htmx
django-template-partials
faker